Split cpp ast info into a new module
